chore: rebase v2 代码，缝合了一下 GitHub Star 提示
