+++ packages/gi-site/src/components/Navbar/SiteNav.tsx
@@ -31,29 +31,7 @@ const styles = {
 
 const SiteNav: React.FunctionComponent<WorkbookBarProps> = props => {
   const { active } = props;
-  const { context, updateContext } = useContext();
   const history = useHistory();
-  const { name } = context;
-
-  const [githubPopVisible, setGithubPopVisible] = React.useState(
-    !localStorage.getItem('GITHUB_POP_CLOSED') || localStorage.getItem('GITHUB_POP_CLOSED') === 'false',
-  );
-
-  const handleCloseGithubPopover = () => {
-    setGithubPopVisible(false);
-    localStorage.setItem('GITHUB_POP_CLOSED', 'true');
-  };
-
-  const handleJumpToGithub = () => {
-    window.open('https://github.com/antvis/G6VP', '_blank');
-    handleCloseGithubPopover();
-  };
-
-  React.useEffect(() => {
-    setTimeout(() => {
-      handleCloseGithubPopover();
-    }, 10000);
-  }, []);
 
   return (
     <header style={styles.container} className="gi-navbar-container">
@@ -73,29 +51,6 @@ const SiteNav: React.FunctionComponent<WorkbookBarProps> = props => {
         <Theme />
         <Github />
         <UserInfo />
-        <Popover
-          title="给个鼓励，加个⭐️吧！"
-          open={githubPopVisible}
-          placement="bottomRight"
-          getPopupContainer={node => node}
-          overlayStyle={{ marginLeft: '20px' }}
-          content={
-            <div style={{ textAlign: 'center' }}>
-              <Button size="small" onClick={handleCloseGithubPopover}>
-                别烦我
-              </Button>
-              <Button size="small" type="primary" style={{ marginLeft: '4px' }} onClick={handleJumpToGithub}>
-                这就去
-              </Button>
-            </div>
-          }
-        >
-          <span className="gi-header-github-icon" style={{ marginLeft: '4px', width: '32px' }}>
-            <a href="http://github.com/antvis/g6vp" target="_blank" rel="noreferrer">
-              <GithubOutlined />
-            </a>
-          </span>
-        </Popover>
       </div>
     </header>
   );
