+++ packages/gi-site/src/components/Navbar/index.less
@@ -62,6 +62,12 @@
   }
 }
 
+.gi-header-github-icon {
+  margin-left: 8px;
+  min-width: 32px;
+  display: inline-flex;
+}
+
 .theme-color-dot {
   width: 8px;
   height: 8px;
